Edits
It's been a long time (several months) since I've worked on this, but it looks like I made it so you get money for picking up (or when they're dropped by the Alien?) extra powerups that you can't hold.

Also took note saying to change the graphics showing which powerups you have active.